Split InputSpec and UrlSpec in 2 files
